fix: in addition to JAVA_OPTS, also generate JAVA_OPTIONS
When creating the Kubernetes/OpenShift resources, Dekorate takes
the `@JvmOptions` annotation and creates the `JAVA_OPTS` environment
variable from it. This environment variable is typically used by
shell scripts that invoke `java` to add cmdline options to it.

However, the `run-java.sh` script, which is used in the Fabric8 S2I
Java image, as well as the Red Hat OpenJDK S2I images, ignores this
environment variable, and instead expects one called `JAVA_OPTIONS`.

Since we can't know in advance which name will be used, we just
generate both, with the same content.